# metal : optimize MoE for large batches